+++ src/MuMu_RichPresence.Tests/Integration/WebScraperTests.cs
@@ -1,11 +1,13 @@
-﻿using Dawn.MuMu.RichPresence.MuMu;
+﻿using Dawn.MuMu.RichPresence.Models;
+using Dawn.MuMu.RichPresence.MuMu;
+using Dawn.MuMu.RichPresence.Scrapers;
 
 namespace MuMu_RichPresence.Tests.Integration;
 
 using System.Collections.Frozen;
 using FluentAssertions;
 
-[TestFixture(TestOf = typeof(PlayStoreWebScraper))]
+[TestFixture(TestOf = typeof(PlayStoreScraper))]
 public class WebScraperTests
 {
     private static readonly FrozenDictionary<string, string> _appPackagesToTitles = new Dictionary<string, string>
@@ -18,7 +20,9 @@ public class WebScraperTests
     public async Task ShouldScrape_Titles(KeyValuePair<string, string> package)
     {
         // Act
-        var packageInfo = await PlayStoreWebScraper.TryGetPackageInfo(package.Key);
+        var session = new MuMuSessionLifetime { AppState = AppState.Focused, PackageName = package.Key, Title = package.Value};
+        
+        var packageInfo = await PackageScraper.TryGetPackageInfo(session);
 
         // Assert
         packageInfo.Should().NotBeNull();