+++ src/MuMu_RichPresence.Tests/Regression/WebScraperTests.cs
@@ -1,4 +1,5 @@
-﻿using Dawn.MuMu.RichPresence.MuMu;
+﻿using Dawn.MuMu.RichPresence.Models;
+using Dawn.MuMu.RichPresence.Scrapers;
 using Polly;
 using Polly.Retry;
 
@@ -6,13 +7,14 @@ namespace MuMu_RichPresence.Tests.Regression;
 
 using FluentAssertions;
 
-[TestFixture(TestOf = typeof(PlayStoreWebScraper))]
+[TestFixture(TestOf = typeof(PlayStoreScraper))]
 public class WebScraperTests
 {
+    private readonly PlayStoreScraper _playStoreScraper = new();
     private static readonly string[] _validPackages = ["com.YoStarEN.Arknights", "com.nexon.bluearchive"];
     private static readonly string[] _invalidPackages = ["com.android.vending", "com.android.browser"];
 
-    private static readonly AsyncRetryPolicy<PlayStoreWebScraper.PlayStorePackageInfo?> _noRetryPolicy = Policy<PlayStoreWebScraper.PlayStorePackageInfo?>
+    private static readonly AsyncRetryPolicy<StorePackageInfo?> _noRetryPolicy = Policy<StorePackageInfo?>
         .Handle<Exception>()
         .WaitAndRetryAsync(1, _ => TimeSpan.FromSeconds(0));
 
@@ -22,7 +24,8 @@ public class WebScraperTests
     public async Task TryGetInfoAsync_WithValidAppPackage_ReturnsValidLink(string packageName)
     {
         // Act
-        var packageInfo = await PlayStoreWebScraper.TryGetPackageInfo(packageName, _noRetryPolicy);
+        var session = new MuMuSessionLifetime { Title = string.Empty,  AppState = AppState.Focused, PackageName = packageName };
+        var packageInfo = await _playStoreScraper.TryGetPackageInfo(session, _noRetryPolicy);
 
         var link = packageInfo?.IconLink;
 
@@ -39,7 +42,8 @@ public class WebScraperTests
     public async Task TryGetInfoAsync_WithInvalidAppPackage_ReturnsNull(string packageName)
     {
         // Act
-        var packageInfo = await PlayStoreWebScraper.TryGetPackageInfo(packageName, _noRetryPolicy);
+        var session = new MuMuSessionLifetime { Title = string.Empty,  AppState = AppState.Focused, PackageName = packageName };
+        var packageInfo = await _playStoreScraper.TryGetPackageInfo(session, _noRetryPolicy);
 
         // Assert
         packageInfo.Should().BeNull();
@@ -49,7 +53,8 @@ public class WebScraperTests
     public async Task TryGetInfoAsync_WithEmptyAppPackage_ReturnsNull()
     {
         // Act
-        var packageInfo = await PlayStoreWebScraper.TryGetPackageInfo(string.Empty, _noRetryPolicy);
+        var session = new MuMuSessionLifetime { Title = string.Empty,  AppState = AppState.Focused, PackageName = string.Empty };
+        var packageInfo = await _playStoreScraper.TryGetPackageInfo(session, _noRetryPolicy);
             
         // Assert
         packageInfo.Should().BeNull();