+++ src/MuMu_RichPresence/Scrapers/Models/BiliGameEntry.cs
@@ -1,0 +1,4 @@
+﻿namespace Dawn.MuMu.RichPresence.Scrapers.Models;
+
+// There are a lot more properties, but we only need the ones listed below
+public record BiliGameEntry(int GameBaseId,  string GameName, string? Icon);