+++ src/MuMu_RichPresence/Scrapers/StorePackageInfo.cs
@@ -1,0 +1,3 @@
+﻿namespace Dawn.MuMu.RichPresence.Scrapers;
+
+public record StorePackageInfo(string IconLink, string Title);