+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 🦺 Changed checking the log file for updates from every `100ms` to every `1 second`
 - ⚡️Added opening the log file as a tray option if ExtendedLogging is enabled
 - 🦺 Bugfix: Toggling the Rich Presence from the tray would not re-enable the presence under certain circumstances
+- ⚡️Added `BiliGame` as an additional app store to get metadata from!
+  - Metadata like the app icon or the app name
 
 ## [2.0.0] / 2025-12-23
 - 🦺 Bugfix: Fixed Rich Presences displaying old game durations in some cases